{{cookiecutter.project_name}}/backend/requirements.txt
Flask
Flask-JWT-Extended
flask-marshmallow
Jinja2
loguru
greenlet
gunicorn
marshmallow
pytest
pytest-cov
pytest-mock
python-dotenv
requests
urllib3
Werkzeug

{%- if cookiecutter.use_db == "yes" %}
Flask-Migrate
Flask-SQLAlchemy
marshmallow
marshmallow-sqlalchemy
psycopg2-binary
SQLAlchemy
{%- endif %}